Zephyro/requirements.txt
Flask
requests
python-dotenv
Flask-Compress
urllib3
flask_cors
torch
numpy
pandas
scikit-learn
